[image] Add option to require all loaded images be trusted
This allows a client to ensure every loaded image is verified,
to protect against for instance tampered modules or initial ram disks.

This change automatically marks DER and PEM images as trusted, as
this would otherwise create a verification loop where new signatures
would be needed to verify the existing PEM/DER signatures. It also
marks embedded images as automatically trusted due to verification
of them being normally not possible.